feat(shadertoy iMouse): Removes iMouse as built-in uniform
This adds the ability for the shadertoy plugin to define props for when iMouse is detected. As modV
doesn't have a vec4 control method, it's up to the user to define a way to translate controls to
vec4 uniforms. The shadertoy plugin is an example of this.

BREAKING CHANGE: Removes iMouse as a built-in uniform

re #170