fix(grab-canvas): Use low resolution for faster performance
Instead of using a 200x200 px canvas to get the data from (160000 values), we now use a 50x50 px
canvas. We are also not sending the buffer to the worker anymore, but the correct data. This means
we don't have to convert the data in the worker back to Uintarray, as it already has the correct
type.