fix(tweens): fixes timings for the first value in tween creation
Changes to using value based animation in animejs instead of keyframes. Keyframes seem to add an
extra frame internally that we can't change.

BREAKING CHANGE: Tweens created in presets prior to this release will have their timings modified if
the BPM triggers an update or the user updates a tween control