feat(ui-focus): adds store module to track focused ui element
To accomodate group deletion via keypress we require some place to track the currently focused UI
item. This presents complications with double-clicking to add modules to a focused group, so this
breaks the existing functionality slightly. Groups are no longer infinitely focusable. If a module
or other UI element becomes focused, the group will blur and double clicking a module in the gallery
will not add the module to the group.

re #486